+++ spec/active_interaction/integration/hash_interaction_spec.rb
@@ -19,10 +19,10 @@ describe HashInteraction do
   include_context 'interactions'
   it_behaves_like 'an interaction', :hash, -> { {} }
 
-  context 'with options[:a]' do
+  context 'with inputs[:a]' do
     let(:a) { { x: {} } }
 
-    before { options.merge!(a: a) }
+    before { inputs.merge!(a: a) }
 
     it 'returns the correct value for :a' do
       expect(result[:a]).to eq a.symbolize_keys